refactor(profile): supprime les anciens composants de gestion du profil
- Retire `ProfileHeaderComponent`, `ProfileInfoComponent` et `ProfileSecurityComponent` ainsi que leurs dépendances associées
- Réorganise les fichiers en externalisant le template du nouveau `ProfileInfo` dans un fichier HTML dédié
- Supprime les imports inutiles et nettoie la structure pour une meilleure maintenabilité
